refactor : UserAuthErrorcode 클래스 이동
