+++ src/main/kotlin/com/example/cmc_be/common/contorller/HealthController.kt
@@ -1,7 +1,7 @@
 package com.example.cmc_be.common.contorller
 
 import com.backend.cmcapi.common.annotation.ApiErrorCodeExample
-import com.example.cmc_be.user.exeption.UserAuthErrorCode
+import com.example.cmc_be.domain.user.exeption.UserAuthErrorCode
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.GetMapping
